+++ examples/resources/fmc_ftd_manual_nat_rule/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_ftd_manual_nat_rule.example "<ftd_nat_policy_id>,<id>"